requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==1.9
django-admin-bootstrapped==2.5.7
idna==2.8
lxml==4.3.0
Pillow==5.4.1
pytesseract==0.2.6
pytz==2018.9
requests==2.21.0
tika==1.19
urllib3==1.24.1
